feat: resolve task modules
